chore: move equals function to reflect test
